requirements.txt
streamlit
cv2
pandas
numpy
pillow